+++ config/routes.rb
@@ -42,7 +42,7 @@ Rails.application.routes.draw do
 
     scope "/pages" do
       get "/" => "pages#index", as: :form_pages
-      post "/" => "page_list#update"
+      post "/" => "forms#mark_pages_section_completed"
       post "/move-page" => "pages#move_page", as: :move_page
 
       scope "/new" do
